Pass SnapshotHierarchy to WatcherRegistry
So each registry can decide what information to use from
the snapshot hierarchy.
