--- subprojects/core/src/main/java/org/gradle/internal/vfs/impl/WatchRootUtil.java
@@ -1,48 +1,0 @@
-/*
- * Copyright 2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.internal.vfs.impl;
-
-import java.nio.file.Path;
-import java.util.Comparator;
-import java.util.HashSet;
-import java.util.Set;
-
-public class WatchRootUtil {
-    /**
-     * Filters out directories whose ancestor is also among the watched directories.
-     */
-    public static Set<Path> resolveRootsToWatch(Set<Path> directories) {
-        Set<Path> roots = new HashSet<>();
-        directories.stream()
-            .sorted(Comparator.comparingInt(Path::getNameCount))
-            .filter(path -> {
-                Path parent = path;
-                while (true) {
-                    parent = parent.getParent();
-                    if (parent == null) {
-                        break;
-                    }
-                    if (roots.contains(parent)) {
-                        return false;
-                    }
-                }
-                return true;
-            })
-            .forEach(roots::add);
-        return roots;
-    }
-}
